pyproject.toml
[project]
name = "pixiv-analysis"
version = "0.1.0"
description = "Pixiv account analysis collector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pixivpy3>=3.7.0",
  "python-dateutil>=2.9.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
